fix get geolocation coordinates from XMP string BUGS: 458754 BUGS: 461522 BUGS: 461633 FIXED-IN: 7.9.0

+++ core/libs/metadataengine/engine/metaengine_gps.cpp
@@ -1026,7 +1026,7 @@ bool MetaEngine::convertFromGPSCoordinateString(const QString& gpsString, double
 
         *degrees  = parts[0].toLong();
         *degrees += parts[1].toLong() / 60.0;
-        *degrees += parts[2].toLong() / 3600.0;
+        *degrees += parts[2].toDouble() / 3600.0;
 
         if ((directionReference == 'W') || (directionReference == 'S'))
         {
